Remove assumption that server is run from root
Previously, reading the deployment_sha file assumed that the server is
always run via `node server.js` from the root directory of the service
and would crash if this is not the case.
In some environments, we may not want to invoke the service in that way,
for example because we want to provide an absolute path to the server.js
file. This change supports that use-case.