Add one-click install script for project
The featureService is not super easy to set up: one must install
postgres, postgis, populate the database, install the node service, etc.
As such, it's worth automating the process to make it easier to set up a
new instance of the featureService; that's what the new `install.sh`
script does.

The script installs postgres and the featureService app on the same host
so that we don't have to manage multiple servers for the database and
application layers. The featureService is exposed over port 80 so that
postgres's port can be hidden in the firewall and we don't have to worry
about securing the database.