Add option to filter any query by name
This is generally useful, e.g. imagine querying for all features in a
bounding box or at a particular point but while already having an idea
of the names of features that we want to retrieve. Additionally, the
feature will be used in the Fortis dashboard for implementing
auto-suggest in the place search pane.